services/vector-api-fastapi/pyproject.toml
[project]
name = "vector-api-fastapi"
version = "0.1.0"
description = "Local DINOv3 vector service for ZRT trademark search."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.133.0",
  "pydantic-settings==2.12.0",
  "uvicorn[standard]==0.38.0",
  "transformers==5.2.0",
  "torch==2.8.0",
  "pillow==12.0.0",
  "safetensors==0.6.2",
]

[dependency-groups]
dev = [
  "httpx==0.28.1",
  "pytest==8.4.2",
  "ruff==0.14.6",
]

[tool.uv]
package = false
